feat(DB-267): image 도메인 리펙토링
