+++ dongsan-core/domain-rdb/src/main/java/com/dongsan/rdb/domains/image/Image.java
@@ -1,0 +1,30 @@
+package com.dongsan.rdb.domains.image;
+
+import com.dongsan.rdb.domains.common.BaseEntity;
+import jakarta.persistence.*;
+
+@Entity
+@Table(name = "image")
+public class Image extends BaseEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false)
+    private String url;
+
+    protected Image() {
+    }
+
+    public Image(String url) {
+        this.url = url;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getUrl() {
+        return url;
+    }
+}
