+++ dongsan-core/domain-rdb/src/main/java/com/dongsan/rdb/domains/image/ImageJpaRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ImageJpaRepository extends JpaRepository<ImageEntity, Long> {
+public interface ImageJpaRepository extends JpaRepository<Image, Long> {
 }
