+++ dongsan-core/domain-rdb/src/main/java/com/dongsan/rdb/domains/image/ImageRepository.java
@@ -1,0 +1,12 @@
+package com.dongsan.rdb.domains.image;
+
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface ImageRepository {
+    Optional<Image> findById(Long imageId);
+
+    Long save(String url);
+}
