+++ dongsan-core/domain-rdb/src/main/java/com/dongsan/rdb/domains/image/ImageService.java
@@ -1,0 +1,26 @@
+package com.dongsan.rdb.domains.image;
+
+import com.dongsan.core.support.error.CoreErrorCode;
+import com.dongsan.core.support.error.CoreException;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional
+public class ImageService {
+    private final ImageRepository imageRepository;
+
+    public ImageService(ImageRepository imageRepository) {
+        this.imageRepository = imageRepository;
+    }
+
+    public Long save(String url) {
+        return imageRepository.save(url);
+    }
+
+    public Image getImage(Long imageId) {
+        return imageRepository.findById(imageId)
+                .orElseThrow(() -> new CoreException(CoreErrorCode.IMAGE_NOT_EXISTS));
+    }
+
+}
