--- dongsan-core/src/main/java/com/dongsan/core/domains/image/Image.java
@@ -1,7 +1,0 @@
-package com.dongsan.core.domains.image;
-
-public record Image(
-	Long imageId,
-	String url
-) {
-}
