--- dongsan-domain/domain-rdb/src/testFixtures/java/fixture/ImageFixture.java
@@ -1,15 +1,0 @@
-package fixture;
-
-import com.dongsan.rdb.domains.image.ImageEntity;
-
-public class ImageFixture {
-	private static final String URL = "https://test.com";
-
-	public static ImageEntity createImage(String url) {
-		return new ImageEntity(url);
-	}
-
-	public static ImageEntity createImage() {
-		return new ImageEntity(URL);
-	}
-}
